qrToSVG を worker 化してみた
