+++ src/QRCode.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import clsx from "clsx"
-  import qrcode from "qrcode-generator"
+  import { qrToSVG } from "./util/qrToSVG"
 
   export let data: string
   export let width: number
@@ -8,43 +8,16 @@
   export { className as class }
   let className: string = ""
 
-  $: ({ size, d } = qrToSVG(data))
-
-  function qrToSVG(data: string): {
-    size: number
-    d: string
-  } {
-    const typeAuto = 0
-    const qr = qrcode(typeAuto, "M")
-    qr.addData(data)
-    qr.make()
-
-    const count = qr.getModuleCount()
-    const margin = 1
-
-    const size = margin + count + margin
-
-    let d = ""
-    for (let row = 0; row < count; row += 1) {
-      for (let col = 0; col < count; col += 1) {
-        if (qr.isDark(row, col)) {
-          d += `M${margin + col},${margin + row}l1,0 0,1 -1,0 0,-1z `
-        }
-      }
-    }
-
-    return {
-      size,
-      d,
-    }
-  }
+  $: svg$ = qrToSVG(data)
 </script>
 
-<svg
-  class={clsx("bg-white text-black", className)}
-  viewBox="0 0 {size} {size}"
-  {width}
-  {height}
->
-  <path stroke="transparent" fill="currentColor" {d} />
-</svg>
+{#await svg$ then { size, d }}
+  <svg
+    class={clsx("bg-white text-black", className)}
+    viewBox="0 0 {size} {size}"
+    {width}
+    {height}
+  >
+    <path stroke="transparent" fill="currentColor" {d} />
+  </svg>
+{/await}
