+++ src/util/qrToSVG.ts
@@ -1,0 +1,5 @@
+import { wrap } from "comlink"
+import type { qrToSVG as qrToSVGType } from "./qrToSVG.worker"
+import QRToSVGWorker from "./qrToSVG.worker?worker&inline"
+
+export const qrToSVG = wrap<typeof qrToSVGType>(new QRToSVGWorker())
